AutoMapper package removed + Web project renamed to Api

+++ src/TimeHacker.Domain.Contracts/Entities/ScheduleSnapshots/ScheduledTask.cs
@@ -15,7 +15,7 @@ namespace TimeHacker.Domain.Contracts.Entities.ScheduleSnapshots
         public bool IsFixed { get; set; }
         public string Name { get; set; }
         public string? Description { get; set; }
-        public uint Priority { get; set; }
+        public byte Priority { get; set; }
         public bool IsCompleted { get; set; } = false;
         public TimeSpan Start { get; set; }
         public TimeSpan End { get; set; }
